financial_education_app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.7.4

openai==1.14.0
agno==0.1.0

langchain-core==0.1.52
langchain-openai==0.1.4
langchain-community>=0.2.0
sentence-transformers>=2.2.0

chromadb==0.4.24

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
